Create check_mailbox_size.sh
A way to manage the problems related to the mailbox when it is full, on a Ubuntu system. I match the mailbox_size_limit in /etc/postfix/main.cf with the current size of the mailbox. When the mailbox is close to be full, this script will empty it. Supports Performance Data.